Add many changes from the katzenpost monorepo
Here we delete a bunch of stuff from the ref implementation such
that the go compiler is able to follow all the dependencies
and build correctly with the C components.

The alternative to this would be to build a .a or .so and install that somewhere
and then include that via CFLAGS so that the go compiler can build against it
and the relevant .h header files.

The motivation to do it this way instead is that it makes it possible
to install this via go install; if the requisite CGO_CFLAGS_ALLOW var is set.